+++ Examples/Qen.cs
@@ -24,8 +24,7 @@ namespace PLGL.Examples
 
             //Structural
             SetLetters();
-            SetSigma();
-            SetPaths();
+            SetStructure();
 
             //Lexemes
             SetInflections();
@@ -39,7 +38,7 @@ namespace PLGL.Examples
             lang.Options.Pathing = LanguageOptions.LetterPathing.Inclusion;
             lang.Options.MemorizeWords = false;
             lang.Options.SigmaSkewMin = 0.8;
-            lang.Options.SigmaSkewMax = 2;
+            lang.Options.SigmaSkewMax = 1.5;
             lang.Options.SeedOffset = 2;
 
             lang.Options.AllowAutomaticCasing = true;
@@ -65,26 +64,26 @@ namespace PLGL.Examples
         }
         private void SetDeconstructEvents()
         {
-            lang.Deconstruct += (lg, current, left, right) => lg.DECONSTRUCT_MergeBlocks(current, left, right, "PUNCTUATION", "LETTERS", "LETTERS", "\'", "LETTERS");
-            lang.Deconstruct += (lg, current, left, right) => lg.DECONSTRUCT_MergeBlocks(current, left, right, "PUNCTUATION", "NUMBERS", "NUMBERS", ".", "NUMBERS");
-            lang.Deconstruct += (lg, current, left, right) => lg.DECONSTRUCT_MergeBlocks(current, left, right, "PUNCTUATION", "NUMBERS", "NUMBERS", ",", "NUMBERS");
-            lang.Deconstruct += (lg, current, left, right) => lg.DECONSTRUCT_ChangeFilter(current, left, right, "PUNCTUATION", "LETTERS", "LETTERS", "-", "COMPOUND");
-            lang.Deconstruct += (lg, current, left, right) => lg.DECONSTRUCT_MergeBlocks(current, left, right, "LETTERS", "ESCAPE", "ESCAPE", "ESCAPE");
-            lang.Deconstruct += (lg, current, left, right) => lg.DECONSTRUCT_ContainWithin(current, left, right, "FLAGSOPEN", "FLAGSCLOSE", "FLAGS");
+            lang.OnDeconstruct += (lg, current, left, right) => lg.DECONSTRUCT_MergeBlocks(current, left, right, "PUNCTUATION", "LETTERS", "LETTERS", "\'", "LETTERS");
+            lang.OnDeconstruct += (lg, current, left, right) => lg.DECONSTRUCT_MergeBlocks(current, left, right, "PUNCTUATION", "NUMBERS", "NUMBERS", ".", "NUMBERS");
+            lang.OnDeconstruct += (lg, current, left, right) => lg.DECONSTRUCT_MergeBlocks(current, left, right, "PUNCTUATION", "NUMBERS", "NUMBERS", ",", "NUMBERS");
+            lang.OnDeconstruct += (lg, current, left, right) => lg.DECONSTRUCT_ChangeFilter(current, left, right, "PUNCTUATION", "LETTERS", "LETTERS", "-", "COMPOUND");
+            lang.OnDeconstruct += (lg, current, left, right) => lg.DECONSTRUCT_MergeBlocks(current, left, right, "LETTERS", "ESCAPE", "ESCAPE", "ESCAPE");
+            lang.OnDeconstruct += (lg, current, left, right) => lg.DECONSTRUCT_ContainWithin(current, left, right, "FLAGSOPEN", "FLAGSCLOSE", "FLAGS");
         }
         private void SetConstructEvents()
         {
-            lang.Construct += (lg, word) => lg.CONSTRUCT_KeepAsIs(word, "UNDEFINED");
-            lang.Construct += (lg, word) => lg.CONSTRUCT_KeepAsIs(word, "DELIMITER");
-            lang.Construct += (lg, word) => lg.CONSTRUCT_Hide(word, "COMPOUND");
+            lang.OnConstruct += (lg, word) => lg.CONSTRUCT_KeepAsIs(word, "UNDEFINED");
+            lang.OnConstruct += (lg, word) => lg.CONSTRUCT_KeepAsIs(word, "DELIMITER");
+            lang.OnConstruct += (lg, word) => lg.CONSTRUCT_Hide(word, "COMPOUND");
 
-            lang.Construct += (lg, word) => lg.CONSTRUCT_Generate(word, "LETTERS");
+            lang.OnConstruct += (lg, word) => lg.CONSTRUCT_Generate(word, "LETTERS");
 
             SetPunctuation();
             SetFlagging();
             SetNumbers();
 
-            lang.Construct += (lg, word) => lg.CONSTRUCT_Within(word, "ESCAPE", 1, 2);
+            lang.OnConstruct += (lg, word) => lg.CONSTRUCT_Within(word, "ESCAPE", 1, 2);
         }
         #endregion
 
@@ -143,9 +142,11 @@ namespace PLGL.Examples
             lang.Alphabet.AddConsonant('ŝ', ('ŝ', 'Ŝ'), 10); //Sh ŝ (ship)
             lang.Alphabet.AddConsonant('Þ', ('Þ', 'þ'), 7); //Th þ (thatch)
             lang.Alphabet.AddConsonant('ż', ('ż', 'Ż'), 1); //Ezh ʒ (azure)
-
+        }
+        private void SetStructure()
+        {
             lang.Structure.AddGroup('V', "Vowels", ('a', 1.0), ('e', 1.0), ('u', 1.0), ('o', 1.0), ('u', 1.0),
-                                                           ('ä', 1.0), ('ë', 1.0), ('ï', 1.0), ('ö', 1.0), ('ü', 1.0));
+                                               ('ä', 1.0), ('ë', 1.0), ('ï', 1.0), ('ö', 1.0), ('ü', 1.0));
             lang.Structure.AddGroup('o', "Vowels (short)", ('a', 1.0), ('e', 1.0), ('u', 1.0), ('o', 1.0), ('u', 1.0));
             lang.Structure.AddGroup('O', "Vowels (long)", ('ä', 1.0), ('ë', 1.0), ('ï', 1.0), ('ö', 1.0), ('ü', 1.0));
 
@@ -158,8 +159,6 @@ namespace PLGL.Examples
             lang.Structure.AddGroup('A', "Approximant", ('w', 1.0), ('y', 1.0), ('h', 1.0));
             lang.Structure.AddGroup('R', "R/L", ('r', 66), ('l', 33));
 
-            lang.Structure.AddSyllable("VR", 1.0);
-
             lang.Structure.AddSyllable("VP", 1.0);
             lang.Structure.AddSyllable("NVP", 2.0);
             lang.Structure.AddSyllable("SpVN", 3.0);
@@ -171,306 +170,6 @@ namespace PLGL.Examples
             lang.Structure.AddSyllable("VN", 1.0);
             lang.Structure.AddSyllable("AV", 1.0);
             lang.Structure.AddSyllable("on", 0.5);
-        }
-        private void SetSigma()
-        {
-            lang.Structure.AddSigma("C", "V", "", new SigmaPath() { SelectionWeight = 10.0, StartingWeight = 1.0, EndingWeight = 1.0, LastConsonantWeight = 1.5 });
-            lang.Structure.AddSigma("", "V", "", new SigmaPath() { SelectionWeight = 3.0, StartingWeight = 1.0, EndingWeight = 1.0, LastConsonantWeight = 1.5 });
-            lang.Structure.AddSigma("CC", "V", "", new SigmaPath() { SelectionWeight = 2.0, StartingWeight = 1.0, EndingWeight = 1.0, LastConsonantWeight = 1.5 });
-            lang.Structure.AddSigma("C", "V", "C", new SigmaPath() { SelectionWeight = 10.0, StartingWeight = 1.0, EndingWeight = 1.0, LastConsonantWeight = 1.5 });
-
-
-        }
-
-        /// <summary>
-        /// Aa, Ee, Ii, Oo, Uu (short)
-        /// Ää, Ëë, Ïï, Öö, Üü (long)
-        /// Bb Pp, Dd Tt, Gg Kk
-        /// Mm Nn, Ŋŋ, Rr Ll
-        /// Ff, Vv, Ss, Zz
-        /// Ww, Yy, Hh, Qq, Ŝŝ, þÞ, Żż
-        /// </summary>
-        private void SetPaths()
-        {
-            //Idea: Add leter group class, and make sigmas match letter grouping instead of standard C/V.
-            //Imagine these letter groups: "short vowels" (v), "long vowels" (V), "fricatives" (f), plosives (p), nasal (n), r / l (r)
-            //Instead of CVC, it could be frvp (possible letters generated: flik, vrep,), 
-
-            #region Vowels (short)
-            lang.Structure.AddLetterPath('a', WordPosition.Last, SigmaPosition.Coda, ('ŋ', 1.0), ('ż', 1.0));
-            lang.Structure.AddLetterPath('e', WordPosition.Last, SigmaPosition.Coda, ('ŋ', 1.0), ('ż', 1.0));
-            lang.Structure.AddLetterPath('i', WordPosition.Last, SigmaPosition.Coda, ('ŋ', 1.0), ('ż', 1.0));
-            lang.Structure.AddLetterPath('o', WordPosition.Last, SigmaPosition.Coda, ('ŋ', 1.0), ('ż', 1.0));
-            lang.Structure.AddLetterPath('u', WordPosition.Last, SigmaPosition.Coda, ('ŋ', 1.0), ('ż', 1.0));
-
-            lang.Structure.AddLetterPath('ä', WordPosition.Last, SigmaPosition.Coda, ('ŋ', 1.0), ('ż', 1.0));
-            lang.Structure.AddLetterPath('ë', WordPosition.Last, SigmaPosition.Coda, ('ŋ', 1.0), ('ż', 1.0));
-            lang.Structure.AddLetterPath('ï', WordPosition.Last, SigmaPosition.Coda, ('ŋ', 1.0), ('ż', 1.0));
-            lang.Structure.AddLetterPath('ö', WordPosition.Last, SigmaPosition.Coda, ('ŋ', 1.0), ('ż', 1.0));
-            lang.Structure.AddLetterPath('ü', WordPosition.Last, SigmaPosition.Coda, ('ŋ', 1.0), ('ż', 1.0));
-
-            lang.Structure.AddLetterPath('a', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-
-            lang.Structure.AddLetterPath('e', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-
-            lang.Structure.AddLetterPath('i', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-
-            lang.Structure.AddLetterPath('o', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-
-            lang.Structure.AddLetterPath('u', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-            #endregion
-
-            #region Vowels (long)
-            lang.Structure.AddLetterPath('ä', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-
-            lang.Structure.AddLetterPath('ë', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-
-            lang.Structure.AddLetterPath('ï', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-
-            lang.Structure.AddLetterPath('ö', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-
-            lang.Structure.AddLetterPath('ü', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-            #endregion
-
-            #region Plosive (b, p, d, t, g, k)
-            lang.Structure.AddLetterPath('b', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 10.0), ('p', 0.0), ('d', 0.0), ('t', 0.0), ('g', 0.0), ('k', 0.0),
-                ('m', 0.0), ('n', 0.0), ('ŋ', 0.0), ('r', 5.0), ('l', 3.0),
-                ('f', 0.0), ('v', 0.0), ('s', 5.0), ('ŝ', 2.0), ('Þ', 0.0), ('z', 0.0),
-                ('w', 0.0), ('y', 0.0), ('h', 0.0), ('q', 0.0));
-
-            lang.Structure.AddLetterPath('p', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('ŋ', 0.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-
-            lang.Structure.AddLetterPath('d', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 0.0), ('p', 0.0), ('d', 5.0), ('t', 0.0), ('g', 0.0), ('k', 0.0),
-                ('m', 7.0), ('n', 7.0), ('ŋ', 0.0), ('r', 7.0), ('l', 7.0),
-                ('f', 1.0), ('v', 1.0), ('s', 8.0), ('ŝ', 0.0), ('Þ', 0.0), ('z', 2.0),
-                ('w', 0.0), ('y', 0.0), ('h', 0.0), ('q', 0.0));
-
-            lang.Structure.AddLetterPath('t', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('ŋ', 0.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-
-            lang.Structure.AddLetterPath('g', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('ŋ', 0.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-
-            lang.Structure.AddLetterPath('k', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('ŋ', 0.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-            #endregion
-
-            #region Nasal (m, n, ŋ, r, l)
-            lang.Structure.AddLetterPath('m', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('n', 7.0), ('ŋ', 0.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-
-            lang.Structure.AddLetterPath('n', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('ŋ', 0.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-
-            lang.Structure.AddLetterPath('ŋ', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('ŋ', 0.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-
-            lang.Structure.AddLetterPath('r', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('ŋ', 0.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-
-            lang.Structure.AddLetterPath('l', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('ŋ', 0.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-            #endregion
-
-            #region Fricatives (f, v, s, z)
-            lang.Structure.AddLetterPath('f', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('ŋ', 0.0), ('r', 7.0), ('l', 7.0),
-                ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-
-            lang.Structure.AddLetterPath('v', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('ŋ', 0.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-
-            lang.Structure.AddLetterPath('s', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('ŋ', 0.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-
-            lang.Structure.AddLetterPath('z', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('ŋ', 0.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-            #endregion
-
-            #region Approximants / Fricatives (w, y, h, q, ŝ, Þ, ż)
-            lang.Structure.AddLetterPath('w', WordPosition.Any, SigmaPosition.Any,
-                ('a', 10.0), ('e', 5.0), ('i', 2.0), ('o', 2.0), ('u', 1.0),
-                ('ä', 8.0), ('ë', 0.5), ('ï', 5.0), ('ö', 1.0), ('ü', 0.5),
-                ('b', 0.0), ('p', 0.0), ('d', 0.0), ('t', 0.0), ('g', 0.0), ('k', 0.0),
-                ('m', 0.0), ('n', 0.0), ('ŋ', 0.0), ('r', 1.0), ('l', 0.0),
-                ('f', 0.0), ('v', 0.0), ('s', 0.0), ('ŝ', 0.0), ('Þ', 0.0), ('z', 0.0),
-                ('w', 0.0), ('y', 0.0), ('h', 0.0), ('q', 0.0));
-
-            lang.Structure.AddLetterPath('y', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('ŋ', 0.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0),
-                ('w', 7.0), ('y', 7.0), ('h', 7.0), ('q', 7.0));
-
-            lang.Structure.AddLetterPath('h', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('ŋ', 0.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0));
-
-            lang.Structure.AddLetterPath('q', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('ŋ', 0.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0));
-
-            lang.Structure.AddLetterPath('ŝ', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('ŋ', 0.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0));
-
-            lang.Structure.AddLetterPath('Þ', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('ŋ', 0.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0));
-
-            lang.Structure.AddLetterPath('ż', WordPosition.Any, SigmaPosition.Any,
-                ('a', 2.0), ('e', 1.0), ('i', 10.0), ('o', 0.25), ('u', 7.0),
-                ('ä', 7.0), ('ë', 7.0), ('ï', 7.0), ('ö', 7.0), ('ü', 7.0),
-                ('b', 7.0), ('p', 7.0), ('d', 7.0), ('t', 7.0), ('g', 7.0), ('k', 7.0),
-                ('m', 7.0), ('n', 7.0), ('ŋ', 0.0), ('r', 7.0), ('l', 7.0),
-                ('f', 7.0), ('v', 7.0), ('s', 7.0), ('ŝ', 7.0), ('Þ', 7.0), ('z', 7.0));
-            #endregion
-
         }
         #endregion
 
@@ -494,7 +193,7 @@ namespace PLGL.Examples
         #region Punctuation
         private void SetPunctuation()
         {
-            lang.Construct += (lg, word) => lang.Punctuation.Process(lg, word, "PUNCTUATION");
+            lang.OnConstruct += (lg, word) => lang.Punctuation.Process(lg, word, "PUNCTUATION");
 
             lang.Punctuation.Add(".", (w) => { return "৹"; });
             lang.Punctuation.Add(",", (w) => { return ","; });
@@ -507,24 +206,23 @@ namespace PLGL.Examples
         #region Flagging
         public void SetFlagging()
         {
-            lang.Construct += (lg, word) => lang.Flags.Process(lg, word, "FLAGS");
+            lang.OnConstruct += (lg, word) => lang.Flags.Process(lg, word, "FLAGS");
 
             lang.Flags.Add("<HIDE", lang.Flags.ACTION_HideLeft);
             lang.Flags.Add("HIDE>", lang.Flags.ACTION_HideRight);
             lang.Flags.Add("<HIDE>", lang.Flags.ACTION_HideAdjacents);
             lang.Flags.Add("NOGEN", lang.Flags.ACTION_NoGenerate);
-            lang.Flags.Add("NOPLURAL", lang.Flags.ACTION_NoAffixes);
+            lang.Flags.Add("NOAFFIX", lang.Flags.ACTION_NoAffixes);
         }
         #endregion
 
         public void SetNumbers()
         {
-            lang.Construct += (lg, word) => lang.Numbers.Process(lg, word, "NUMBERS", '-', Char.MinValue);
+            lang.OnConstruct += (lg, word) => lang.Numbers.Process(lg, word, "NUMBERS", '-', Char.MinValue);
             lang.Numbers.Add(('0', '○', ""),
                 ('1', '•', ""), ('2', '›', ""), ('3', '△', ""),
                 ('4', '◇', ""), ('5', '⨰', ""), ('6', '∓', ""),
-                ('7', '⪦', ""), ('8', '⋇', ""), ('9', '⨳', ""),
-                ('A', 'A', ""), ('B', 'B', ""));
+                ('7', '⪦', ""), ('8', '⋇', ""), ('9', '⨳', ""));
         }
     }
 }